# openapiDescription.yaml
openapi: 3.0.3
info:
  title: Meeting & Transcription Platform API
  description: >
    API pour la plateforme web et mobile de gestion, enregistrement et transcription de réunions.
    Elle regroupe les endpoints des microservices Meeting et Transcription.
  version: 1.0.0
  contact:
    name: Projet transcript Réunions
    email: contact@contact.local
servers:
  - url: http://localhost:8080
    description: Environnement local (Quarkus / Docker)
  - url: https://api.reunion-platform.com
    description: Environnement de production
...
paths:
...
  /api/meeting:
    post:
      tags: [Meeting]
      summary: Créer une nouvelle réunion
      operationId: createMeeting
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MeetingCreateRequest'
      responses:
        '201':
          description: Réunion créée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Meeting'
...
  /api/meeting/{id_meeting}:
    get:
      tags: [Meeting]
      summary: Obtenir les détails d’une réunion spécifique
      operationId: getMeetingById
      parameters:
        - name: id_meeting
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Détails de la réunion
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Meeting'
...
components:
  schemas:
    Meeting:
      type: object
      properties:
        id_meeting:
          type: integer
        title:
          type: string
        date:
          type: string
          format: date-time
        participants:
          type: array
          items:
            $ref: '#/components/schemas/Participant'
        status:
          type: string
          enum: [Planifiée, En cours, Terminée]
    MeetingCreateRequest:
      type: object
      properties:
        title:
          type: string
        date:
          type: string
          format: date-time
        participants:
          type: array
          items:
            $ref: '#/components/schemas/Participant'
    Participant:
      type: object
      properties:
        id_participant:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
...